Refactored ProjectOverviewTab.js to a functional Typescript component.  This process seems to have fixed the issue.
